Split ACPI Functions from SMP functions. Add ACPI Table detection for FADT & HPET (as we need better timers)
